requirements.txt
openai
pandas
pydantic
python-dotenv
openenv-core
fastapi
uvicorn